+++ app/src/main/java/com/aabdelaal/droos/data/source/local/LocalDataSource.kt
@@ -12,9 +12,9 @@ interface LocalDataSource {
     fun getTeachers(): Result<LiveData<List<TeacherInfo>>>
     fun getTeachersByStatus(isActive: Boolean): Result<LiveData<List<TeacherInfo>>>
     suspend fun saveTeacherInfo(teacherInfo: TeacherInfo): Long
-    suspend fun getTeacherInfoById(id: Int): Result<TeacherInfo>
+    suspend fun getTeacherInfoById(id: Long): Result<TeacherInfo>
     suspend fun deleteAllTeachers()
-    suspend fun deleteTeacherInfo(id: Int)
+    suspend fun deleteTeacherInfo(id: Long)
 
     fun getSubjects(): Result<LiveData<List<Subject>>>
     suspend fun saveSubject(subject: Subject): Long
