+++ app/src/main/java/com/aabdelaal/droos/data/source/remote/models/RemoteSubject.kt
@@ -14,7 +14,7 @@ data class RemoteSubject(
     @DocumentId
     var remoteID: String?,
 
-    val id: Int = 0
+    val id: Long = 0
 
 ) {
 }
